Complete A-Packages; Create folders for different packages
