+++ src/main/java/duke/taskType/Deadline.java
@@ -1,3 +1,5 @@
+package duke.taskType;
+
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
